Move validator rewards to runtime
