+++ near/src/runtime.rs
@@ -1,4 +1,3 @@
-use std::cmp::max;
 use std::collections::{HashMap, HashSet};
 use std::convert::{TryFrom, TryInto};
 use std::fs::File;
@@ -30,12 +29,11 @@ use near_primitives::views::{
     AccessKeyInfoView, CallResult, QueryError, QueryResponse, ViewStateResult,
 };
 use near_store::{
-    get_access_key_raw, get_account, set_account, StorageError, Store, StoreUpdate, Trie,
-    TrieUpdate, WrappedTrieChanges, COL_STATE,
+    get_access_key_raw, Store, StoreUpdate, Trie, TrieUpdate, WrappedTrieChanges, COL_STATE,
 };
 use node_runtime::adapter::ViewRuntimeAdapter;
 use node_runtime::state_viewer::TrieViewer;
-use node_runtime::{ApplyState, Runtime, StateRecord};
+use node_runtime::{ApplyState, Runtime, StateRecord, ValidatorAccountsUpdate};
 
 use crate::config::GenesisConfig;
 use crate::shard_tracker::{account_id_to_shard_id, ShardTracker};
@@ -124,68 +122,6 @@ impl NightshadeRuntime {
             epoch_manager,
             shard_tracker,
         }
-    }
-
-    /// Iterates over validator accounts in the given shard and updates their accounts to return stake
-    /// and allocate rewards.
-    fn update_validator_accounts(
-        &self,
-        shard_id: ShardId,
-        block_hash: &CryptoHash,
-        last_validator_proposals: &[ValidatorStake],
-        state_update: &mut TrieUpdate,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        let mut epoch_manager = self.epoch_manager.write().expect(POISONED_LOCK_ERR);
-        let (stake_info, validator_reward) = epoch_manager.compute_stake_return_info(block_hash)?;
-        let account_to_stake =
-            last_validator_proposals.iter().fold(HashMap::new(), |mut acc, v| {
-                acc.insert(v.account_id.clone(), v.amount);
-                acc
-            });
-
-        for (account_id, max_of_stakes) in stake_info {
-            if self.account_id_to_shard_id(&account_id) == shard_id {
-                let account: Option<Account> = get_account(state_update, &account_id)?;
-                if let Some(mut account) = account {
-                    if let Some(reward) = validator_reward.get(&account_id) {
-                        debug!(target: "runtime", "account {} adding reward {} to stake {}", account_id, reward, account.locked);
-                        account.locked += *reward;
-                    }
-
-                    debug!(target: "runtime",
-                           "account {} stake {} max_of_stakes: {}",
-                           account_id, account.locked, max_of_stakes
-                    );
-                    assert!(
-                        account.locked >= max_of_stakes,
-                        "FATAL: staking invariant does not hold. Account stake {} is less than maximum of stakes {} in the past three epochs",
-                        account.locked,
-                        max_of_stakes
-                    );
-                    let last_stake = *account_to_stake.get(&account_id).unwrap_or(&0);
-                    let return_stake = account.locked - max(max_of_stakes, last_stake);
-                    debug!(target: "runtime", "account {} return stake {}", account_id, return_stake);
-                    account.locked -= return_stake;
-                    account.amount += return_stake;
-
-                    set_account(state_update, &account_id, &account);
-                }
-            }
-        }
-        if self.account_id_to_shard_id(&self.genesis_config.protocol_treasury_account) == shard_id {
-            let mut protocol_treasury_account =
-                get_account(state_update, &self.genesis_config.protocol_treasury_account)?.unwrap();
-            protocol_treasury_account.amount +=
-                *validator_reward.get(&self.genesis_config.protocol_treasury_account).unwrap();
-            set_account(
-                state_update,
-                &self.genesis_config.protocol_treasury_account,
-                &protocol_treasury_account,
-            );
-        }
-        state_update.commit();
-
-        Ok(())
     }
 
     fn genesis_state_from_dump(&self) -> (StoreUpdate, Vec<StateRoot>) {
@@ -603,32 +539,44 @@ impl RuntimeAdapter for NightshadeRuntime {
         } else {
             self.trie.clone()
         };
-        let mut state_update = TrieUpdate::new(trie.clone(), state_root.hash);
-        let should_update_account = {
+        let validator_accounts_update = {
             let mut epoch_manager = self.epoch_manager.write().expect(POISONED_LOCK_ERR);
             debug!(target: "runtime",
                 "block index: {}, is next_block_epoch_start {}",
                 block_index,
                 epoch_manager.is_next_block_epoch_start(prev_block_hash).unwrap()
             );
-            epoch_manager.is_next_block_epoch_start(prev_block_hash)?
+            if epoch_manager.is_next_block_epoch_start(prev_block_hash)? {
+                let (stake_info, validator_reward) =
+                    epoch_manager.compute_stake_return_info(prev_block_hash)?;
+                let stake_info = stake_info
+                    .into_iter()
+                    .filter(|(account_id, _)| self.account_id_to_shard_id(account_id) == shard_id)
+                    .collect();
+                let validator_rewards = validator_reward
+                    .into_iter()
+                    .filter(|(account_id, _)| self.account_id_to_shard_id(account_id) == shard_id)
+                    .collect();
+                let last_proposals = last_validator_proposals
+                    .iter()
+                    .filter(|v| self.account_id_to_shard_id(&v.account_id) == shard_id)
+                    .fold(HashMap::new(), |mut acc, v| {
+                        acc.insert(v.account_id.clone(), v.amount);
+                        acc
+                    });
+                Some(ValidatorAccountsUpdate {
+                    stake_info,
+                    validator_rewards,
+                    last_proposals,
+                    protocol_treasury_account_id: Some(
+                        self.genesis_config.protocol_treasury_account.clone(),
+                    )
+                    .filter(|account_id| self.account_id_to_shard_id(account_id) == shard_id),
+                })
+            } else {
+                None
+            }
         };
-
-        // If we are starting to apply 1st block in the new epoch.
-        if should_update_account {
-            self.update_validator_accounts(
-                shard_id,
-                prev_block_hash,
-                last_validator_proposals,
-                &mut state_update,
-            )
-            .map_err(|e| {
-                if let Some(e) = e.downcast_ref::<StorageError>() {
-                    panic!(e.to_string())
-                }
-                Error::from(ErrorKind::ValidatorError(e.to_string()))
-            })?;
-        }
 
         let apply_state = ApplyState {
             block_index,
@@ -639,7 +587,14 @@ impl RuntimeAdapter for NightshadeRuntime {
 
         let apply_result = self
             .runtime
-            .apply(state_update, &apply_state, &receipts, &transactions)
+            .apply(
+                trie.clone(),
+                state_root.hash,
+                &validator_accounts_update,
+                &apply_state,
+                &receipts,
+                &transactions,
+            )
             .map_err(|e| match e {
                 InvalidTxErrorOrStorageError::InvalidTxError(_) => ErrorKind::InvalidTransactions,
                 InvalidTxErrorOrStorageError::StorageError(_) => {
