+++ runtime/runtime/src/lib.rs
@@ -5,7 +5,7 @@ extern crate serde_derive;
 #[macro_use]
 extern crate lazy_static;
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::convert::TryInto;
 
 use borsh::BorshSerialize;
@@ -31,13 +31,14 @@ use near_primitives::utils::{
 use near_runtime_fees::RuntimeFeesConfig;
 use near_store::{
     get, get_access_key, get_account, get_receipt, get_received_data, set, set_access_key,
-    set_account, set_code, set_receipt, set_received_data, StorageError, StoreUpdate, TrieChanges,
-    TrieUpdate,
+    set_account, set_code, set_receipt, set_received_data, StorageError, StoreUpdate, Trie,
+    TrieChanges, TrieUpdate,
 };
 use near_vm_logic::types::PromiseResult;
 use near_vm_logic::ReturnData;
 
 use crate::actions::*;
+use crate::balance_checker::check_balance;
 use crate::config::{
     exec_fee, safe_add_balance, safe_add_gas, safe_gas_to_balance, total_deposit, total_exec_fees,
     total_prepaid_gas, total_send_fees, RuntimeConfig,
@@ -47,9 +48,12 @@ use near_primitives::errors::{
     ActionError, ExecutionError, InvalidAccessKeyError, InvalidTxError,
     InvalidTxErrorOrStorageError,
 };
+use std::cmp::max;
+use std::sync::Arc;
 
 mod actions;
 pub mod adapter;
+mod balance_checker;
 pub mod cache;
 pub mod config;
 pub mod ext;
@@ -69,6 +73,18 @@ pub struct ApplyState {
     pub gas_price: Balance,
     /// A block timestamp
     pub block_timestamp: u64,
+}
+
+/// Contains information to update validators accounts at the first block of a new epoch.
+pub struct ValidatorAccountsUpdate {
+    /// Maximum stake across last 3 epochs.
+    pub stake_info: HashMap<AccountId, Balance>,
+    /// Rewards to distribute to validators.
+    pub validator_rewards: HashMap<AccountId, Balance>,
+    /// Stake proposals from the last chunk.
+    pub last_proposals: HashMap<AccountId, Balance>,
+    /// The ID of the protocol treasure account if it belongs to the current shard.
+    pub protocol_treasury_account_id: Option<AccountId>,
 }
 
 #[derive(Debug)]
@@ -836,8 +852,59 @@ impl Runtime {
         Ok(None)
     }
 
+    /// Iterates over validator in the current shard and updates their accounts to return stake
+    /// and allocate rewards. Also updates protocol treasure account if it belongs to the current
+    /// shard.
+    fn update_validator_accounts(
+        &self,
+        state_update: &mut TrieUpdate,
+        validator_accounts_update: &ValidatorAccountsUpdate,
+    ) -> Result<(), StorageError> {
+        for (account_id, max_of_stakes) in &validator_accounts_update.stake_info {
+            if let Some(mut account) = get_account(state_update, account_id)? {
+                if let Some(reward) = validator_accounts_update.validator_rewards.get(account_id) {
+                    debug!(target: "runtime", "account {} adding reward {} to stake {}", account_id, reward, account.locked);
+                    account.locked += *reward;
+                }
+
+                debug!(target: "runtime",
+                       "account {} stake {} max_of_stakes: {}",
+                       account_id, account.locked, max_of_stakes
+                );
+                assert!(
+                    account.locked >= *max_of_stakes,
+                    "FATAL: staking invariant does not hold.\
+                     Account stake {} is less than maximum of stakes {} in the past three epochs",
+                    account.locked,
+                    max_of_stakes
+                );
+                let last_proposal =
+                    *validator_accounts_update.last_proposals.get(account_id).unwrap_or(&0);
+                let return_stake = account.locked - max(*max_of_stakes, last_proposal);
+                debug!(target: "runtime", "account {} return stake {}", account_id, return_stake);
+                account.locked -= return_stake;
+                account.amount += return_stake;
+
+                set_account(state_update, account_id, &account);
+            }
+        }
+
+        if let Some(account_id) = &validator_accounts_update.protocol_treasury_account_id {
+            // If protocol treasury stakes, then the rewards was already distributed above.
+            if !validator_accounts_update.stake_info.contains_key(account_id) {
+                let mut account = get_account(state_update, account_id)?.unwrap();
+                account.amount +=
+                    *validator_accounts_update.validator_rewards.get(account_id).unwrap();
+                set_account(state_update, account_id, &account);
+            }
+        }
+        state_update.commit();
+
+        Ok(())
+    }
     /// Applies new singed transactions and incoming receipts for some chunk/shard on top of
-    /// given root of the state.
+    /// given trie and the given state root.
+    /// If the validator accounts update is provided, updates validators accounts.
     /// All new signed transactions should be valid and already verified by the chunk producer.
     /// If any transaction is invalid, it would return an `InvalidTxError`.
     /// Returns an `ApplyResult` that contains the new state root, trie changes,
@@ -846,13 +913,19 @@ impl Runtime {
     /// receivers) and incoming action receipts.
     pub fn apply(
         &self,
-        mut state_update: TrieUpdate,
+        trie: Arc<Trie>,
+        root: CryptoHash,
+        validator_accounts_update: &Option<ValidatorAccountsUpdate>,
         apply_state: &ApplyState,
         prev_receipts: &[Receipt],
         transactions: &[SignedTransaction],
     ) -> Result<ApplyResult, InvalidTxErrorOrStorageError> {
-        // TODO(#1481): Remove clone after Runtime bug is fixed.
-        let initial_state = state_update.clone();
+        let initial_state = TrieUpdate::new(trie.clone(), root);
+        let mut state_update = TrieUpdate::new(trie.clone(), root);
+
+        if let Some(validator_accounts_update) = validator_accounts_update {
+            self.update_validator_accounts(&mut state_update, validator_accounts_update)?;
+        }
 
         let mut new_receipts = Vec::new();
         let mut validator_proposals = vec![];
@@ -884,9 +957,11 @@ impl Runtime {
             .for_each(|res| tx_result.push(res));
         }
 
-        self.check_balance(
+        check_balance(
+            &self.config.transaction_costs,
             &initial_state,
             &state_update,
+            validator_accounts_update,
             prev_receipts,
             transactions,
             &new_receipts,
@@ -902,150 +977,6 @@ impl Runtime {
             tx_result,
             stats,
         })
-    }
-
-    // TODO: Check for balance overflows
-    // TODO: Fix StorageError for partial states when looking up something that doesn't exist.
-    fn check_balance(
-        &self,
-        initial_state: &TrieUpdate,
-        final_state: &TrieUpdate,
-        prev_receipts: &[Receipt],
-        transactions: &[SignedTransaction],
-        new_receipts: &[Receipt],
-        stats: &ApplyStats,
-    ) -> Result<(), InvalidTxErrorOrStorageError> {
-        // Accounts
-        let all_accounts_ids: HashSet<AccountId> = transactions
-            .iter()
-            .map(|tx| tx.transaction.signer_id.clone())
-            .chain(prev_receipts.iter().map(|r| r.receiver_id.clone()))
-            .collect();
-        let total_accounts_balance = |state| -> Result<u128, StorageError> {
-            Ok(all_accounts_ids
-                .iter()
-                .map(|account_id| {
-                    Ok(get_account(state, account_id)?.map_or(0, |a| a.amount + a.locked))
-                })
-                .collect::<Result<Vec<u128>, StorageError>>()?
-                .into_iter()
-                .sum::<u128>())
-        };
-        let initial_account_balance = total_accounts_balance(&initial_state)?;
-        let final_account_balance = total_accounts_balance(&final_state)?;
-        // Receipts
-        let receipt_cost = |receipt: &Receipt| -> Result<u128, InvalidTxError> {
-            Ok(match &receipt.receipt {
-                ReceiptEnum::Action(action_receipt) => {
-                    let mut total_cost = total_deposit(&action_receipt.actions)?;
-                    if receipt.predecessor_id != system_account() {
-                        let mut total_gas = safe_add_gas(
-                            self.config.transaction_costs.action_receipt_creation_config.exec_fee(),
-                            total_exec_fees(
-                                &self.config.transaction_costs,
-                                &action_receipt.actions,
-                            )?,
-                        )?;
-                        total_gas =
-                            safe_add_gas(total_gas, total_prepaid_gas(&action_receipt.actions)?)?;
-                        let total_gas_cost =
-                            safe_gas_to_balance(action_receipt.gas_price, total_gas)?;
-                        total_cost = safe_add_balance(total_cost, total_gas_cost)?;
-                    }
-                    total_cost
-                }
-                ReceiptEnum::Data(_) => 0,
-            })
-        };
-        let receipts_cost = |receipts: &[Receipt]| {
-            receipts
-                .iter()
-                .map(receipt_cost)
-                .collect::<Result<Vec<u128>, InvalidTxError>>()
-                .expect(OVERFLOW_CHECKED_ERR)
-                .into_iter()
-                .sum::<u128>()
-        };
-        let incoming_receipts_balance = receipts_cost(prev_receipts);
-        let outgoing_receipts_balance = receipts_cost(new_receipts);
-        // Postponed actions receipts. The receipts can be postponed and stored with the receiver's
-        // account ID when the input data is not received yet.
-        // We calculate all potential receipts IDs that might be postponed initially or after the
-        // execution.
-        let all_potential_postponed_receipt_ids: HashSet<(AccountId, CryptoHash)> = prev_receipts
-            .iter()
-            .map(|receipt| {
-                let account_id = &receipt.receiver_id;
-                match &receipt.receipt {
-                    ReceiptEnum::Action(_) => {
-                        Ok(Some((account_id.clone(), receipt.receipt_id.clone())))
-                    }
-                    ReceiptEnum::Data(data_receipt) => {
-                        if let Some(receipt_id) = get(
-                            initial_state,
-                            &key_for_postponed_receipt_id(account_id, &data_receipt.data_id),
-                        )? {
-                            Ok(Some((account_id.clone(), receipt_id)))
-                        } else {
-                            Ok(None)
-                        }
-                    }
-                }
-            })
-            .collect::<Result<Vec<Option<_>>, StorageError>>()?
-            .into_iter()
-            .filter_map(|x| x)
-            .collect();
-
-        let total_postponed_receipts_cost = |state| -> Result<u128, InvalidTxErrorOrStorageError> {
-            Ok(all_potential_postponed_receipt_ids
-                .iter()
-                .map(|(account_id, receipt_id)| {
-                    Ok(get_receipt(state, account_id, &receipt_id)?
-                        .map_or(Ok(0), |r| receipt_cost(&r))?)
-                })
-                .collect::<Result<Vec<u128>, InvalidTxErrorOrStorageError>>()?
-                .into_iter()
-                .sum::<u128>())
-        };
-        let initial_postponed_receipts_balance = total_postponed_receipts_cost(initial_state)?;
-        let final_postponed_receipts_balance = total_postponed_receipts_cost(final_state)?;
-        // Sum it up
-        let initial_balance = initial_account_balance
-            + incoming_receipts_balance
-            + initial_postponed_receipts_balance;
-        let final_balance = final_account_balance
-            + outgoing_receipts_balance
-            + final_postponed_receipts_balance
-            + stats.total_rent_paid
-            + stats.total_validator_reward
-            + stats.total_balance_burnt;
-        if initial_balance != final_balance {
-            panic!(
-                "Runtime Apply initial balance sum {} doesn't match final balance sum {}\n\
-                 \tInitial account balance sum: {}\n\
-                 \tFinal account balance sum: {}\n\
-                 \tIncoming receipts balance sum: {}\n\
-                 \tOutgoing receipts balance sum: {}\n\
-                 \tInitial postponed receipts balance sum: {}\n\
-                 \tFinal postponed receipts balance sum: {}\n\
-                 \t{:?}\n\
-                 \tIncoming Receipts {:#?}\n\
-                 \tOutgoing Receipts {:#?}",
-                initial_balance,
-                final_balance,
-                initial_account_balance,
-                final_account_balance,
-                incoming_receipts_balance,
-                outgoing_receipts_balance,
-                initial_postponed_receipts_balance,
-                final_postponed_receipts_balance,
-                stats,
-                prev_receipts,
-                new_receipts,
-            );
-        }
-        Ok(())
     }
 
     pub fn compute_storage_usage(&self, records: &[StateRecord]) -> HashMap<AccountId, u64> {
@@ -1206,8 +1137,6 @@ mod tests {
     use testlib::runtime_utils::{alice_account, bob_account};
 
     use super::*;
-    use near_crypto::{InMemorySigner, KeyType};
-    use near_primitives::transaction::TransferAction;
 
     #[test]
     fn test_get_and_set_accounts() {
@@ -1235,140 +1164,72 @@ mod tests {
         assert_eq!(test_account, get_res);
     }
 
-    /***********************/
-    /* Balance check tests */
-    /***********************/
+    /***************/
+    /* Apply tests */
+    /***************/
 
     #[test]
-    fn test_check_balance_no_op() {
+    fn test_apply_no_op() {
         let trie = create_trie();
         let root = MerkleHash::default();
-        let initial_state = TrieUpdate::new(trie.clone(), root);
-        let final_state = TrieUpdate::new(trie.clone(), root);
         let runtime = Runtime::new(RuntimeConfig::default());
-        runtime
-            .check_balance(&initial_state, &final_state, &[], &[], &[], &ApplyStats::default())
-            .unwrap();
-    }
 
-    #[test]
-    #[should_panic]
-    fn test_check_balance_unaccounted_refund() {
-        let trie = create_trie();
-        let root = MerkleHash::default();
-        let initial_state = TrieUpdate::new(trie.clone(), root);
-        let final_state = TrieUpdate::new(trie.clone(), root);
-        let runtime = Runtime::new(RuntimeConfig::default());
-        runtime
-            .check_balance(
-                &initial_state,
-                &final_state,
-                &[Receipt::new_refund(&alice_account(), 1000)],
-                &[],
-                &[],
-                &ApplyStats::default(),
-            )
-            .unwrap_or_else(|_| ());
-    }
-
-    #[test]
-    fn test_check_balance_refund() {
-        let trie = create_trie();
-        let root = MerkleHash::default();
         let account_id = alice_account();
 
         let initial_balance = 1_000_000;
-        let refund_balance = 1000;
 
         let mut initial_state = TrieUpdate::new(trie.clone(), root);
         let initial_account = Account::new(initial_balance, hash(&[]), 0);
         set_account(&mut initial_state, &account_id, &initial_account);
-        initial_state.commit();
+        let trie_changes = initial_state.finalize().unwrap();
+        let (store_update, root) = trie_changes.into(trie.clone()).unwrap();
+        store_update.commit().unwrap();
 
-        let mut final_state = TrieUpdate::new(trie.clone(), root);
-        let final_account = Account::new(initial_balance + refund_balance, hash(&[]), 0);
-        set_account(&mut final_state, &account_id, &final_account);
-        final_state.commit();
+        let apply_state =
+            ApplyState { block_index: 0, epoch_length: 3, gas_price: 100, block_timestamp: 100 };
 
-        let runtime = Runtime::new(RuntimeConfig::default());
-        runtime
-            .check_balance(
-                &initial_state,
-                &final_state,
-                &[Receipt::new_refund(&account_id, refund_balance)],
-                &[],
-                &[],
-                &ApplyStats::default(),
-            )
-            .unwrap();
+        runtime.apply(trie, root, &None, &apply_state, &[], &[]).unwrap();
     }
 
     #[test]
-    fn test_check_balance_tx_to_receipt() {
+    fn test_apply_check_balance_validation_rewards() {
         let trie = create_trie();
         let root = MerkleHash::default();
+        let runtime = Runtime::new(RuntimeConfig::default());
+
         let account_id = alice_account();
 
-        let initial_balance = 1_000_000_000;
-        let deposit = 500_000_000;
-        let gas_price = 100;
-        let runtime_config = RuntimeConfig::default();
-        let cfg = &runtime_config.transaction_costs;
-        let exec_gas = cfg.action_receipt_creation_config.exec_fee()
-            + cfg.action_creation_config.transfer_cost.exec_fee();
-        let send_gas = cfg.action_receipt_creation_config.send_fee(false)
-            + cfg.action_creation_config.transfer_cost.send_fee(false);
-        let contract_reward = (send_gas * cfg.burnt_gas_reward.numerator
-            / cfg.burnt_gas_reward.denominator) as Balance
-            * gas_price;
-        let total_validator_reward = send_gas as Balance * gas_price - contract_reward;
+        let initial_balance = 1_000_000;
+        let initial_locked = 500_000;
+        let reward = 10_000_000;
+        let small_refund = 500;
+
         let mut initial_state = TrieUpdate::new(trie.clone(), root);
-        let initial_account = Account::new(initial_balance, hash(&[]), 0);
+        let mut initial_account = Account::new(initial_balance, hash(&[]), 0);
+        initial_account.locked = initial_locked;
         set_account(&mut initial_state, &account_id, &initial_account);
-        initial_state.commit();
+        let trie_changes = initial_state.finalize().unwrap();
+        let (store_update, root) = trie_changes.into(trie.clone()).unwrap();
+        store_update.commit().unwrap();
 
-        let mut final_state = TrieUpdate::new(trie.clone(), root);
-        let final_account = Account::new(
-            initial_balance - (exec_gas + send_gas) as Balance * gas_price - deposit
-                + contract_reward,
-            hash(&[]),
-            0,
-        );
-        set_account(&mut final_state, &account_id, &final_account);
-        final_state.commit();
+        let apply_state =
+            ApplyState { block_index: 0, epoch_length: 3, gas_price: 100, block_timestamp: 100 };
 
-        let signer = InMemorySigner::from_seed(&account_id, KeyType::ED25519, &account_id);
-        let tx = SignedTransaction::send_money(
-            1,
-            account_id.clone(),
-            bob_account(),
-            &signer,
-            deposit,
-            CryptoHash::default(),
-        );
-        let receipt = Receipt {
-            predecessor_id: tx.transaction.signer_id.clone(),
-            receiver_id: tx.transaction.receiver_id.clone(),
-            receipt_id: Default::default(),
-            receipt: ReceiptEnum::Action(ActionReceipt {
-                signer_id: tx.transaction.signer_id.clone(),
-                signer_public_key: tx.transaction.public_key.clone(),
-                gas_price,
-                output_data_receivers: vec![],
-                input_data_ids: vec![],
-                actions: vec![Action::Transfer(TransferAction { deposit })],
-            }),
+        let validator_accounts_update = ValidatorAccountsUpdate {
+            stake_info: vec![(account_id.clone(), initial_locked)].into_iter().collect(),
+            validator_rewards: vec![(account_id.clone(), reward)].into_iter().collect(),
+            last_proposals: Default::default(),
+            protocol_treasury_account_id: None,
         };
 
-        let runtime = Runtime::new(runtime_config);
         runtime
-            .check_balance(
-                &initial_state,
-                &final_state,
+            .apply(
+                trie,
+                root,
+                &Some(validator_accounts_update),
+                &apply_state,
+                &[Receipt::new_refund(&account_id, small_refund)],
                 &[],
-                &[tx],
-                &[receipt],
-                &ApplyStats { total_rent_paid: 0, total_validator_reward, total_balance_burnt: 0 },
             )
             .unwrap();
     }
