+++ test-utils/testlib/src/user/runtime_user.rs
@@ -69,10 +69,9 @@ impl RuntimeUser {
         let mut txs = transactions;
         loop {
             let mut client = self.client.write().expect(POISONED_LOCK_ERR);
-            let state_update = TrieUpdate::new(client.trie.clone(), client.state_root);
             let apply_result = client
                 .runtime
-                .apply(state_update, &apply_state, &receipts, &txs)
+                .apply(client.trie.clone(), client.state_root, &None, &apply_state, &receipts, &txs)
                 .map_err(|e| match e {
                     InvalidTxErrorOrStorageError::InvalidTxError(e) => format!("{}", e),
                     InvalidTxErrorOrStorageError::StorageError(e) => {
